Refactoring (main) - 47
