+++ controllers/contacts/updatePutContact.js
@@ -7,23 +7,23 @@ const { lineBreak } = require("../../service");
 
 //-----------------------------------------------------------------------------
 //* ++++++++++++++++++++++ Схема ВАЛИДАЦИИ Joi +++++++++++++++++++++++++
-const contactSchemaPostPut = Joi.object({
-    name: Joi.string()
-        // .alphanum()
-        .min(3)
-        .max(30)
-        .required(),
+// const contactSchemaPostPut = Joi.object({
+//     name: Joi.string()
+//         // .alphanum()
+//         .min(3)
+//         .max(30)
+//         .required(),
 
-    email: Joi.string()
-        .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net', 'ua', 'org',] } })
-        .required(),
+//     email: Joi.string()
+//         .email({ minDomainSegments: 2, tlds: { allow: ['com', 'net', 'ua', 'org',] } })
+//         .required(),
 
-    phone: Joi.string()
-        // .alphanum()
-        .min(5)
-        .max(14)
-        .required(),
-});
+//     phone: Joi.string()
+//         // .alphanum()
+//         .min(5)
+//         .max(14)
+//         .required(),
+// });
 //* _______________________ Схема ВАЛИДАЦИИ Joi _______________________
 
 
@@ -36,24 +36,24 @@ const updatePutContact = async (req, res, next) => {
         //! ==============================================================
 
         //* +++++++++++++++++++++++ ВАЛИДАЦИЯ Joi +++++++++++++++++++++++++++++
-        const validationResult = contactSchemaPostPut.validate(req.body);
+        // const validationResult = contactSchemaPostPut.validate(req.body);
 
-        if (validationResult.error) {
-            //! ===========================console============================
-            console.log("Ошибка ВАЛИДАЦИИ:".bgRed.black);
-            console.log("");
-            console.log(validationResult.error);
-            lineBreak();
-            console.log("END-->PUT/:id".rainbow); //!
-            //! ==============================================================
-            //! 1 - вариант
-            // return res.status(400).json({ "message": "missing required name field" });
-            //! 2 - вариант
-            // validationResult.error.status = 400
-            // throw validationResult.error
-            //! 3 - вариант
-            return res.status(400).json({ status: validationResult.error.details });
-        }
+        // if (validationResult.error) {
+        //     //! ===========================console============================
+        //     console.log("Ошибка ВАЛИДАЦИИ:".bgRed.black);
+        //     console.log("");
+        //     console.log(validationResult.error);
+        //     lineBreak();
+        //     console.log("END-->PUT/:id".rainbow); //!
+        //     //! ==============================================================
+        //     //! 1 - вариант
+        //     // return res.status(400).json({ "message": "missing required name field" });
+        //     //! 2 - вариант
+        //     // validationResult.error.status = 400
+        //     // throw validationResult.error
+        //     //! 3 - вариант
+        //     return res.status(400).json({ status: validationResult.error.details });
+        // }
         //* __________________________ ВАЛИДАЦИЯ Joi __________________________
 
         const { contactId } = req.params;
